avance_1/requirements.txt
requests==2.31.0
pandas==2.2.3
python-dotenv==1.0.0
matplotlib==3.9.0
openpyxl==3.1.2
psycopg2-binary==2.9.11
sqlalchemy==2.0.23
alembic==1.12.1
streamlit==1.28.1
plotly==5.17.0
loguru==0.7.2
numpy==1.26.0